moved eliminate_join_marks to oracle dialect and minor fixes
