+++ sqlglot/optimizer/__init__.py
@@ -1,9 +1,6 @@
 # ruff: noqa: F401
 
 from sqlglot.optimizer.optimizer import RULES as RULES, optimize as optimize
-from sqlglot.optimizer.eliminate_join_marks import (
-    eliminate_join_marks as eliminate_join_marks,
-)
 from sqlglot.optimizer.scope import (
     Scope as Scope,
     build_scope as build_scope,
