+++ tests/test_optimizer.py
@@ -494,12 +494,12 @@ SELECT :with,WITH :expressions,CTE :this,UNION :this,SELECT :expressions,1,:expr
             pretty=True,
         )
 
-    def test_eliminate_join_marks(self):
-        self.check_file(
-            "eliminate_join_marks",
-            optimizer.eliminate_join_marks,
-            pretty=False,
-        )
+    # def test_eliminate_join_marks(self):
+    #     self.check_file(
+    #         "eliminate_join_marks",
+    #         optimizer.eliminate_join_marks,
+    #         pretty=False,
+    #     )
 
     def test_eliminate_ctes(self):
         self.check_file(
